server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
CouchDB==1.2
Deprecated==1.2.14
dnspython==2.7.0
graphql-core==3.2.5
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
libcst==1.5.0
limits==3.13.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
strawberry-graphql==0.247.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
wrapt==1.16.0

aiolimiter~=1.1.0
slowapi~=0.1.9